backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0
requests>=2.31,<3.0
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0
anthropic>=0.18,<0.19
google-auth>=2.28,<3.0
